Made some refactor and fixed Linux installation
