+++ plugins/python/src/main/kotlin/io/paddle/plugin/python/dependencies/interpretator/GenericInterpreterDownloader.kt
@@ -1,0 +1,6 @@
+package io.paddle.plugin.python.dependencies.interpretator
+
+import io.paddle.project.PaddleProject
+
+internal class GenericInterpreterDownloader(userDefinedVersion: InterpreterVersion, project: PaddleProject) :
+    AbstractInterpreterDownloader(userDefinedVersion, project)
