+++ plugins/python/src/main/kotlin/io/paddle/plugin/python/dependencies/interpretator/PythonPaths.kt
@@ -1,0 +1,6 @@
+package io.paddle.plugin.python.dependencies.interpretator
+
+object PythonPaths {
+    const val PYTHON_DISTRIBUTIONS_BASE_URL = "http://www.python.org/ftp/python/"
+    internal const val LOCAL_PYTHON_DIR_NAME = ".localpython"
+}
